compute_engine/requirements.txt
boto==2.44.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
crcmod==1.7
cryptography==1.7.1
enum34==1.1.6
futures==3.3.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-datastore==1.10.0
google-cloud-speech==1.3.0
google-compute-engine==2.8.13
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2019.3
pyxdg==0.25
requests==2.22.0
rsa==4.0
SecretStorage==2.3.1
six==1.13.0
urllib3==1.25.7
